requirements.txt
praw==7.8.1
python-dotenv==1.1.1
google-generativeai==0.8.5
tqdm==4.67.1
streamlit==1.46.1